created fetch crt data action creator
