+++ 17-AdvancedRedux/src/store/cartSlice.js
@@ -1,5 +1,4 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { uiActions } from "./uiSlice";
 
 const initialState = { items: [], display: false };
 
@@ -31,58 +30,12 @@ const cartSlice = createSlice({
         state.items.splice(index, 1);
       }
     },
+
+    loadCart(state, action) {
+      state.items = action.payload;
+    }
   },
 });
-
-// action creator
-export const sendCartData = (cart) => {
-  // is a function that stores another function, which can be async
-  return async (dispatch) => {
-    dispatch(
-      // update notification status
-      uiActions.status({
-        status: "pending",
-        title: "Pending",
-        message: "Sending request",
-      })
-    );
-
-    const sendRequest = async () => {
-      const response = await fetch(
-        // add reference /card.json to URL
-        "https://advanced-redux-ea825-default-rtdb.firebaseio.com/cart.json",
-        {
-          method: "PUT", // will overwrite existing data with new data
-          body: JSON.stringify(cart), // stringify the data to JSON format
-        }
-      );
-
-      if (!response.ok) {
-        throw new Error("Sending data failed");
-      }
-    };
-
-    try {
-      await sendRequest();
-      dispatch(
-        // no errors, update notification
-        uiActions.status({
-          status: "success",
-          title: "Success",
-          message: "Request sent",
-        })
-      );
-    } catch (error) {
-      dispatch(
-        uiActions.status({
-          status: "error",
-          title: "Error",
-          message: "Failed to send request",
-        })
-      );
-    }
-  };
-};
 
 export const cartActions = cartSlice.actions;
 export default cartSlice.reducer;
